+++ api/lib/routes.js
@@ -1,7 +1,6 @@
 import * as accountRecovery from './application/account-recovery/index.js';
 import * as adminMembers from './application/admin-members/index.js';
 import * as authentication from './application/authentication/index.js';
-import * as authenticationOidc from './application/authentication/oidc/index.js';
 import * as cache from './application/cache/index.js';
 import * as campaignParticipations from './application/campaign-participations/index.js';
 import * as campaigns from './application/campaigns/index.js';
@@ -40,7 +39,6 @@ const routes = [
   accountRecovery,
   adminMembers,
   authentication,
-  authenticationOidc,
   cache,
   campaignParticipations,
   campaigns,
