+++ api/src/identity-access-management/application/oidc-provider/oidc-provider.admin.controller.js
@@ -1,8 +1,8 @@
+import { oidcAuthenticationServiceRegistry } from '../../../../lib/domain/usecases/index.js';
+import { PIX_ADMIN } from '../../../authorization/domain/constants.js';
 import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 import { usecases } from '../../domain/usecases/index.js';
 import * as oidcProviderSerializer from '../../infrastructure/serializers/jsonapi/oidc-identity-providers.serializer.js';
-import { oidcAuthenticationServiceRegistry } from '../../../../lib/domain/usecases/index.js';
-import { PIX_ADMIN } from '../../../authorization/domain/constants.js';
 
 /**
  * @param request
