+++ api/src/identity-access-management/application/oidc-provider/oidc-provider.admin.route.js
@@ -32,4 +32,29 @@ export const oidcProviderAdminRoutes = [
       tags: ['identity-access-management', 'api', 'admin', 'oidc'],
     },
   },
+  {
+    method: 'POST',
+    path: '/api/admin/oidc/user/reconcile',
+    config: {
+      auth: false,
+      validate: {
+        payload: Joi.object({
+          data: Joi.object({
+            attributes: Joi.object({
+              email: Joi.string().email().required(),
+              identity_provider: Joi.string().required(),
+              authentication_key: Joi.string().required(),
+            }),
+            type: Joi.string(),
+          }),
+        }),
+      },
+      handler: (request, h) => oidcProviderAdminController.reconcileUserForAdmin(request, h),
+      notes: [
+        "- Cette route permet d'ajouter le fournisseur d'identité d'où provient l'utilisateur comme méthode de connexion à son compte Pix.\n" +
+          "- Cette action s'effectue après que l'utilisateur se soit identifié auprès de son fournisseur d'identité",
+      ],
+      tags: ['identity-access-management', 'api', 'admin', 'oidc'],
+    },
+  },
 ];
