--- api/tests/acceptance/application/authentication/oidc/reconcile-route-admin-post_test.js
@@ -1,54 +1,0 @@
-import jsonwebtoken from 'jsonwebtoken';
-
-import { authenticationSessionService } from '../../../../../src/identity-access-management/domain/services/authentication-session.service.js';
-import { createServer, databaseBuilder, expect } from '../../../../test-helper.js';
-
-describe('Acceptance | Application | Oidc | Routes', function () {
-  describe('POST /api/admin/oidc/user/reconcile', function () {
-    it('should return 200 HTTP status', async function () {
-      // given
-      const user = databaseBuilder.factory.buildUser.withRawPassword({
-        email: 'eva.poree@example.net',
-        rawPassword: 'pix123',
-      });
-      await databaseBuilder.commit();
-      const server = await createServer();
-
-      const idToken = jsonwebtoken.sign(
-        {
-          given_name: 'Brice',
-          family_name: 'Glace',
-          sub: 'some-user-unique-id',
-        },
-        'secret',
-      );
-      const userAuthenticationKey = await authenticationSessionService.save({
-        sessionContent: { idToken },
-        userInfo: {
-          firstName: 'Brice',
-          lastName: 'Glace',
-          externalIdentityId: 'some-user-unique-id',
-        },
-      });
-
-      // when
-      const response = await server.inject({
-        method: 'POST',
-        url: `/api/admin/oidc/user/reconcile`,
-        payload: {
-          data: {
-            attributes: {
-              identity_provider: 'OIDC_EXAMPLE_NET',
-              authentication_key: userAuthenticationKey,
-              email: user.email,
-            },
-          },
-        },
-      });
-
-      // then
-      expect(response.statusCode).to.equal(200);
-      expect(response.result['access_token']).to.exist;
-    });
-  });
-});
