+++ api/tests/identity-access-management/acceptance/application/oidc-provider.admin.route.test.js
@@ -1,12 +1,16 @@
+import jsonwebtoken from 'jsonwebtoken';
+
+import { authenticationSessionService } from '../../../../src/identity-access-management/domain/services/authentication-session.service.js';
 import {
   createServer,
+  databaseBuilder,
   expect,
   generateValidRequestAuthorizationHeader,
   insertUserWithRoleSuperAdmin,
   knex,
 } from '../../../test-helper.js';
 
-describe('Acceptance | Identity Access Management | Route | Oidc', function () {
+describe('Acceptance | Identity Access Management | Route | Admin | oidc-provider', function () {
   let server;
 
   beforeEach(async function () {
@@ -78,6 +82,53 @@ describe('Acceptance | Identity Access Management | Route | Oidc', function () {
           },
         },
       ]);
+    });
+  });
+
+  describe('POST /api/admin/oidc/user/reconcile', function () {
+    it('returns 200 HTTP status code', async function () {
+      // given
+      const user = databaseBuilder.factory.buildUser.withRawPassword({
+        email: 'eva.poree@example.net',
+        rawPassword: 'pix123',
+      });
+      await databaseBuilder.commit();
+
+      const idToken = jsonwebtoken.sign(
+        {
+          given_name: 'Brice',
+          family_name: 'Glace',
+          sub: 'some-user-unique-id',
+        },
+        'secret',
+      );
+      const userAuthenticationKey = await authenticationSessionService.save({
+        sessionContent: { idToken },
+        userInfo: {
+          firstName: 'Brice',
+          lastName: 'Glace',
+          externalIdentityId: 'some-user-unique-id',
+        },
+      });
+
+      // when
+      const response = await server.inject({
+        method: 'POST',
+        url: `/api/admin/oidc/user/reconcile`,
+        payload: {
+          data: {
+            attributes: {
+              identity_provider: 'OIDC_EXAMPLE_NET',
+              authentication_key: userAuthenticationKey,
+              email: user.email,
+            },
+          },
+        },
+      });
+
+      // then
+      expect(response.statusCode).to.equal(200);
+      expect(response.result['access_token']).to.exist;
     });
   });
 });
